change how `article.hvc.predict` saves predicted labels
Instead of saving a flat text file, where each row
is a string of predicted labels,
instead add the strings directly to a 'pred_labels'
column that is added to the `pandas.DataFrame`
loaded from `extract_csv_path`.
The function now returns a `pred_csv_path`
pointing to the .csv saved with the added
'pred_labels' column.

This ensures predicted labels are in the same row
as the ground truth annotation, instead of
hoping that the order of lines in the flat text file
corresponds to the rows that get used from a separately
loaded csv by the `score` function.
Which is what I was doing before and it turned out
I was doing it wrong -- I forgot to filter the
loaded csv by the 'split' column.

Also rewrote the function to avoid fancy array juggling.
We now just iterate through the ``DataFrame`` once,
loading feature files, predicting from them, and
directly appending the predicted labels to a list,
instead of concatenating all the arrays from
feature files and then splitting them back up afterwards.
The refactored approach is more readable and foolproof --
no need to make sure we split everything back up
correctly.