fix how hvc.score handles empty strings in pred_labels
when a cell in a column contains empty strings,
those are saved as null values, and then loaded
in again as np.nan

to handle this we do a list comprehension on
the values loaded from the pred_labels column,
keeping any strings as is but converting anything
else to an empty string, ''.
This is a good enough solution for this specific case.