BUG: fix how 'semi-automated cleaning' works in `hvc.resegment`
logic that finds onsets and offsets
should have been
"greater than or equal to" first onset and
"less than or equal" last offset, not
"greater than first onset and less than last offset",
since that omitted the first and last valid segment.

Actually ended up using "1 ms before first onset or
1 ms after last offset".
This still in some (rare-ish) cases does not recover
first onset / last offset in ground truth data
but that's the whole point of "semi-automated cleaning".

For BFs after re-running there is a drop in syllable error rate,
but it looks like TweetyNet still ends up performing
as well as it w/~10 minutes of data for BFs.

This doesn't affect canary results where it's plain as day
that a model that does segmenting is needed

- fix bug in hvc/run_learncurve.py
- update segment_error .csv files for BFs
  after re-running hvc/run_learncurve.py
- update segment_error_across_birds.hvc.csv for BFs
- update segment_error csvs for individual canaries
- update segment_error_across_birds.hvc.csv for canaries
- rerun + tweak figure code in src/scripts/hvc/hvc-figure.ipynb
- update figure files in doc/article/figures/mainfig_tweetynet_v_svm
- rename src/scripts/hvc -> src/scripts/tweetynet-v-svm